Exposes the id of the widget
This is important when you are resizing the widgets. Depending on the widget name is the template that you would like to invoke. If you do not know the widget that is being resized then how will you know the inner html resizing algorithm that you need to use. Either way, this is an attribute gridstack already exposes.